% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMSE_Optimal_BW.R
\name{AMSE_Optimal_BW}
\alias{AMSE_Optimal_BW}
\title{AMSE-Optimal Bandwidth Selector for Q-LATE Estimator}
\usage{
AMSE_Optimal_BW(df, qlist)
}
\arguments{
\item{df}{Data.frame. For now, need the following columns:
Y (outcome variable), T (continuous treatment),
R (running variable centered at 0), Z (cutoff indicator),
ZR (cutoff indicator x running variable),
R2 (running variable^2),
and ZR2 (cutoff indicator x running variable^2)}

\item{qlist}{Vector of quantiles to estimate QLATE at. Must be evenly spaced}
}
\description{
AMSE-Optimal Bandwidth Selector for Q-LATE Estimator
}
