% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContinuousRD-package.R
\docType{package}
\name{ContinuousRD-package}
\alias{ContinuousRD}
\alias{ContinuousRD-package}
\title{ContinuousRD: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
