% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QLATE_bc_se.R
\name{QLATE_bc_se}
\alias{QLATE_bc_se}
\title{Implements Quantile-LATE from Dong, Lee, and Gou (2021)}
\usage{
QLATE_bc_se(df, qlist)
}
\arguments{
\item{df}{Data.frame. For now, need the following columns:
Y (outcome variable), T (continuous treatment),
R (running variable centered at 0), Z (cutoff indicator),
ZR (cutoff indicator x running variable),
R2 (running variable^2),
and ZR2 (cutoff indicator x running variable^2)}

\item{qlist}{Vector of quantiles to estimate QLATE at}
}
\description{
Implements the QLATE estimate proposed in Dong, Lee, and Gou and uses a
bias-corrected standard error
}
