% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMSE_Optimal_BW.R
\name{AMSE_Optimal_BW}
\alias{AMSE_Optimal_BW}
\title{AMSE-Optimal Bandwidth Selector for Q-LATE Estimator}
\usage{
AMSE_Optimal_BW(df, qlist, yname, tname, rname, c = 0, verbose = TRUE)
}
\arguments{
\item{df}{Data.frame.}

\item{qlist}{Vector of quantiles to estimate QLATE at. Must be evenly spaced}

\item{yname}{Character. Variable name for outcome variable}

\item{tname}{Character. Variable name for continuous treatment}

\item{rname}{Character. Variable name for running variable.}

\item{c}{Numeric. Cutoff for running variable. Default is 0.}

\item{verbose}{Logical. If TRUE, then print results to the console}
}
\description{
AMSE-Optimal Bandwidth Selector for Q-LATE Estimator
}
