% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QLATE_bc_se.R
\name{QLATE_bc_se}
\alias{QLATE_bc_se}
\title{Implements Quantile-LATE from Dong, Lee, and Gou (2021)}
\usage{
QLATE_bc_se(df, qlist, yname, tname, rname, c = 0, verbose = TRUE)
}
\arguments{
\item{df}{Data.frame.}

\item{qlist}{Vector of quantiles to estimate QLATE at. Must be evenly spaced}

\item{yname}{Character. Variable name for outcome variable}

\item{tname}{Character. Variable name for continuous treatment}

\item{rname}{Character. Variable name for running variable.}

\item{c}{Numeric. Cutoff for running variable. Default is 0.}

\item{verbose}{Logical. If TRUE, then print results to the console}
}
\description{
Implements the QLATE estimate proposed in Dong, Lee, and Gou and uses a
bias-corrected standard error
}
